requirements.txt
# Core dependencies
pydantic>=2.0.0
pyyaml>=6.0
jinja2>=3.1.0
requests>=2.31.0
python-dotenv>=1.0.0

# Retry logic
tenacity>=8.2.0

# CLI (optional, for better UX)
click>=8.1.0

# Development
pytest>=7.4.0
